requirements.txt
fastapi==0.63.0
humanize==3.4.1
python-dotenv==0.17.0
python-json-logger==2.0.1
requests==2.25.1
typing-extensions==3.7.4.3
uvicorn==0.13.4
